move common method to utils
